# docs/swagger.yaml
basePath: /api
definitions:
...
  router.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      teamId:
        type: integer
      username:
        type: string
    type: object
  router.DictEntityResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  router.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  router.ListUsersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/router.UserResponse'
        type: array
    type: object
...
  router.UpdateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      teamId:
        type: integer
      username:
        type: string
    type: object
  router.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/router.DictEntityResponse'
        type: array
      teamId:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Incident management system API
  title: Fixik API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with optional limit and offset
      parameters:
      - default: 100
        description: Limit of items
        in: query
        name: limit
        required: true
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update user
      tags:
      - users
schemes:
- http
swagger: "2.0"